+++ src/nvim/eval/fs.c
@@ -20,6 +20,7 @@
 #include "nvim/eval/fs.h"
 #include "nvim/eval/typval.h"
 #include "nvim/eval/userfunc.h"
+#include "nvim/eval/vars.h"
 #include "nvim/eval/window.h"
 #include "nvim/ex_cmds.h"
 #include "nvim/ex_docmd.h"
