+++ src/nvim/os/shell.c
@@ -11,8 +11,8 @@
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
 #include "nvim/errors.h"
-#include "nvim/eval.h"
 #include "nvim/eval/typval_defs.h"
+#include "nvim/eval/vars.h"
 #include "nvim/event/defs.h"
 #include "nvim/event/libuv_proc.h"
 #include "nvim/event/loop.h"
