+++ src/nvim/os/signal.c
@@ -9,7 +9,7 @@
 #include "nvim/autocmd.h"
 #include "nvim/autocmd_defs.h"
 #include "nvim/buffer_defs.h"
-#include "nvim/eval.h"
+#include "nvim/eval/vars.h"
 #include "nvim/event/defs.h"
 #include "nvim/event/signal.h"
 #include "nvim/ex_cmds2.h"
