+++ src/nvim/syntax.c
@@ -16,7 +16,6 @@
 #include "nvim/cmdexpand_defs.h"
 #include "nvim/drawscreen.h"
 #include "nvim/errors.h"
-#include "nvim/eval.h"
 #include "nvim/eval/typval_defs.h"
 #include "nvim/eval/vars.h"
 #include "nvim/ex_cmds_defs.h"
